Use uint64 in proto for pipeline members

+++ rocAL/source/pipeline/pipeline_serializer.cpp
@@ -41,11 +41,11 @@ void PipelineSerializer::serialize_to_file(const std::string& file_path) {
 }
 
 void PipelineSerializer::serialize_pipeline_config(size_t num_threads, size_t batch_size, int device_id, RocalMemType device_type, size_t prefetch_queue_depth) {
-    _pipeline_proto.set_num_threads(num_threads);
-    _pipeline_proto.set_batch_size(batch_size);
+    _pipeline_proto.set_num_threads(static_cast<uint64_t>(num_threads));
+    _pipeline_proto.set_batch_size(static_cast<uint64_t>(batch_size));
     _pipeline_proto.set_device_id(device_id);
     _pipeline_proto.set_rocal_cpu(device_type == RocalMemType::HOST);
-    _pipeline_proto.set_prefetch_queue_depth(prefetch_queue_depth);
+    _pipeline_proto.set_prefetch_queue_depth(static_cast<uint64_t>(prefetch_queue_depth));
 }
 
 void set_tensor_proto(rocal_proto::InputOutput *in_out_proto, Tensor *tensor, bool is_input) {
